Enemy AI Functional/Shooting Fixed/GameOver Functional
In this version I had implemented another Shooting system for the Player, with new Methods within. As well as the Enemy has now an AI which Patrols, Detects, Chases and Attacks the Player. When the Player's HP is equal to 0 is Game Over, this UI features a restart button to retry the game.

Lots of changes