UI GAME OVER
In this version an UI Game Over was made in the canvas so whenever the Player's HP is equal to 0 this menu is shown.

In the Canvas a Text Mesh Pro has been created with no Text so the ProjectileGun Script has access to this element making an Ammo Display.

Also an Attack Point (Empty Object) has been parented to the gun which allows the laser prefab be instantiate in this point no matter the Player's Movement.

In addition some variables has been setup in the Inspector  due to problem originated by the branch change. For the Enemy the Sight Range is set to 10 and the Attack Range is set to 5 so the Enemy has range to Patrolling and Attack when the Player has entered into the Sight Range. For the Player in the ProjectileGun Script the variables were not assigned so they have been assigned in this version. The Shoot Force is 150. Magazine Size is 20.