UI_GunProp_LaserThicker
In this branch the UI has been changed to an Original Art work delivered by the Art Department.
In this manner, the gun has also been replaced by an Original Design, and the maps have been assigned in the Inspector, changing the Shader mode to Autodesk since from here is possible to assign a Roughness map.

The thickness of the laser is increased in this version, covering more area per damage.

Another change is the value of the Speed Rotation, it is now 150 since in previous versions this movement felt bulky and rough, which impacted directly on the gameplay. With this change, the movement is smoother and faster.