HPBar_WinCondition_SFX
HP Bar has been implemented in the project, the script controls 10 square images in the UI, every 10 HP taken as result of damage, one bar will be disabled.

Sound Effects are implemented with a script that manages the audios, and instantiate the script, in specific the method required, in each case.

An enemy counter has been done to create a list on how many enemies are left, and when the number of enemies equals 0, it is Game Over.

Scenes have been created for the Main Menu, and Game Over, as well as the models for the enemies have been imported.